OK I think I got better API for this
I simply create two posess, and we optinally in production code go to the pose  the one we actually go to (with negative x direction wrt that target pose) before going to the ultimate target pose. This way it handles poses that aren't perfectly oriented towards the base link frames.

Whew, now we can test on the physical robot, I guess ? Then simulator?